First Blueprint Implementation "Auth": Basic views and template system without Auth strategy
