+++ app/__init__.py
@@ -3,7 +3,7 @@ from flask import Flask
 from flask_bootstrap import Bootstrap
 
 from .config import Config
-
+from .auth import auth
 
 def create_app():
     '''retutns an instance of Flask Class'''
@@ -11,5 +11,7 @@ def create_app():
     bootstrap = Bootstrap(app)
 
     app.config.from_object(Config)
+
+    app.register_blueprint(auth)
 
     return app
