+++ app/auth/__init__.py
@@ -1,0 +1,6 @@
+'''Init Auth Blueprint'''
+from flask import Blueprint
+
+auth = Blueprint('auth', __name__, url_prefix='/auth')
+
+from . import views
