--- app/test/test_base.py
@@ -1,50 +1,0 @@
-from flask_testing import TestCase
-from flask import current_app, url_for
-from main import app
-
-
-class MainTest(TestCase):
-    def create_app(self):
-        app.config['TESTING'] = True
-        app.config['WTF_CSRF_ENABLED'] = False
-        return app
-
-    def test_app_exist(self):
-        '''test if app exist'''
-        self.assertIsNotNone(current_app)
-
-
-    def test_app_in_test_mode(self):
-        '''test app config is TESTING'''
-        self.assertTrue(current_app.config['TESTING'])
-
-    def test_index_redirect(self):
-        '''test if app redirects from index to pomodoro_time'''
-        response = self.client.get(url_for('index'))
-        self.assert_redirects(response, url_for('pomodoro_time'))
-
-    def test_pomodoro_time_get(self):
-        '''test status 200 on get pomodoro_time'''
-        response = self.client.get(url_for('pomodoro_time'))
-        self.assert200(response)
-
-    def test_signup_post(self):
-        '''test post form and redirect on signup'''
-        fake_form = {
-            'username': 'fake',
-            'email': 'fake@email.com',
-            'password': 'password-fake'
-        }
-
-        response = self.client.post(url_for('signup'), data=fake_form)
-        self.assert_redirects(response, url_for('index'))
-
-    def test_login_post(self):
-        '''test post form and redirect on login'''
-        fake_form = {
-            'username': 'fake',
-            'password': 'password-fake'
-        }
-
-        response = self.client.post(url_for('login'), data=fake_form)
-        self.assert_redirects(response, url_for('index'))
