+++ requirements.txt
@@ -6,6 +6,7 @@ firebase-admin==4.5.0
 flask-login==0.4.1
 wtforms-validators==1.0.0
 email-validator==1.1.2
+blinker==1.4
